fix: finished modularizing sections
