+++ src/createIcons.js
@@ -1,18 +1,18 @@
-
+import * as THREE from 'three';
 import { TextGeometry } from "TextGeometry";
 import { FontLoader } from "FontLoader";
-import { curvePlanes } from "./utils/geometryUtils";
+import { curvePlanes } from "./utils/geometryUtils.js";
 
-const yPos = 42;
-const zPos = 8.8;
+const yPos = 39;
+const zPos = 8.85;
 
 export function createIcons(scene, textures) {
-    createCenteredIcon(scene, textures.aboutMeTexture, -15, "About Me", yPos, zPos);
-    createCenteredIcon(scene, textures.contactTexture, 15, "Contact", yPos, zPos);
-    createCenteredIcon(scene, textures.experienceTexture, 5, "Experience", yPos, zPos);
-    createCenteredIcon(scene, textures.projectTexture, -5, "Projects", yPos, zPos);
+    createCenteredIcon(textures.aboutMeTexture, -15, "About Me");
+    createCenteredIcon(textures.contactTexture, 15, "Contact");
+    createCenteredIcon(textures.experienceTexture, 5, "Experience");
+    createCenteredIcon(textures.projectTexture, -5, "Projects");
 
-    function createCenteredIcon(texture, positionX, name) {
+    function createCenteredIcon(texture, xPos, name) {
         const planeGeometry = curvePlanes(12.5, 8.5);
         const planeMaterial = new THREE.MeshStandardMaterial({
             map: texture,
@@ -21,13 +21,13 @@ export function createIcons(scene, textures) {
             transparent: true,
         });
         const plane = new THREE.Mesh(planeGeometry, planeMaterial);
-        plane.position.set(positionX, yPos + 2, zPos + 0.05);
+        plane.position.set(xPos, yPos + 2, zPos + 0.05);
         plane.userData = { type: 'icon', text: name };
         scene.add(plane);
-        addIconText(positionX, name);
+        addIconText(xPos, name);
     }
 
-    function addIconText(positionX, name){
+    function addIconText(xPos, name){
         const fontLoader = new FontLoader();
         fontLoader.load("https://unpkg.com/three@0.166.1/examples/fonts/gentilis_bold.typeface.json", function(font) {
             const textGeometry = new TextGeometry(name, {
@@ -42,7 +42,7 @@ export function createIcons(scene, textures) {
             const textMesh = new THREE.Mesh(textGeometry, textMaterial);
             textGeometry.computeBoundingBox();
             const textWidth = textGeometry.boundingBox.max.x - textGeometry.boundingBox.min.x;
-            textMesh.position.set(positionX - textWidth / 2, yPos - 2, zPos * 2 - 0.7);
+            textMesh.position.set(xPos - textWidth / 2, yPos - 2, zPos * 2 - 0.7);
             scene.add(textMesh);
         });
     }
